+++ src/Trulioo.Client.V1/Model/Communication.cs
@@ -19,5 +19,9 @@
         /// Email address of the individual to be verified
         /// </summary>
         public string EmailAddress { get; set; }
+        /// <summary>
+        /// Additional Phone/Mobile Number of the individual to be verified
+        /// </summary>
+        public string Telephone2 { get; set; }
     }
 }